Imported Upstream version 2.3.2

+++ src/common/slurm_jobacct_gather.c
@@ -10,7 +10,7 @@
  *  Copyright (C) 2005 Hewlett-Packard Development Company, L.P.
  *
  *  This file is part of SLURM, a resource management program.
- *  For details, see <https://computing.llnl.gov/linux/slurm/>.
+ *  For details, see <http://www.schedmd.com/slurmdocs/>.
  *  Please also read the included file: DISCLAIMER.
  *
  *  SLURM is free software; you can redistribute it and/or modify it under
@@ -84,7 +84,7 @@ typedef struct slurm_jobacct_gather_ops {
 	void (*jobacct_gather_change_poll)   (uint16_t frequency);
 	void (*jobacct_gather_suspend_poll)  ();
 	void (*jobacct_gather_resume_poll)   ();
-	int (*jobacct_gather_set_proctrack_container_id)(uint32_t id);
+	int (*jobacct_gather_set_proctrack_container_id)(uint64_t id);
 	int (*jobacct_gather_add_task) (pid_t pid, jobacct_id_t *jobacct_id);
 	jobacctinfo_t *(*jobacct_gather_stat_task)(pid_t pid);
 	jobacctinfo_t *(*jobacct_gather_remove_task)(pid_t pid);
@@ -478,7 +478,7 @@ extern void jobacct_gather_g_resume_poll()
 	return;
 }
 
-extern int jobacct_gather_g_set_proctrack_container_id(uint32_t id)
+extern int jobacct_gather_g_set_proctrack_container_id(uint64_t id)
 {
 	int retval = SLURM_SUCCESS;
 	if (_slurm_jobacct_gather_init() < 0)
