+++ src/slurmd/slurmstepd/slurmstepd_job.h
@@ -9,7 +9,7 @@
  *  CODE-OCEC-09-009. All rights reserved.
  *
  *  This file is part of SLURM, a resource management program.
- *  For details, see <https://computing.llnl.gov/linux/slurm/>.
+ *  For details, see <http://www.schedmd.com/slurmdocs/>.
  *  Please also read the included file: DISCLAIMER.
  *
  *  SLURM is free software; you can redistribute it and/or modify it under
@@ -190,7 +190,7 @@ typedef struct slurmd_job {
 	uint16_t       multi_prog;
 	uint16_t       overcommit;
 	env_t          *envtp;
-	uint32_t       cont_id;
+	uint64_t       cont_id;
 
 	char          *batchdir;
 	jobacctinfo_t *jobacct;
